requirements.txt
tensorboard==2.18.0
tensorboardX==2.6.2.2
pytorch3d==0.7.4
easydict==1.13
thop==0.1.1.post2209072238
ninja==1.11.1.1
h5py==3.11.0
matplotlib==3.9.2
numpy==1.26.4
open3d==0.18.0
pyYAML==6.0.1
scipy==1.14.1
timm==1.0.9
tqdm==4.66.4
trimesh==4.4.9
scikit-image==0.24.0
transformers==4.46.2
cuml-cu11==24.10.0